pyproject.toml
[tool.poetry]
name = "siri-ask-gpt"
version = "0.1.0"
description = ""
authors = ["Reliable Magician <petr.b.lavrov@gmail.com>"]
readme = "README.md"
packages = [{ include = "siri_ask_gpt" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "*"
uvicorn = "*"
openai = "*"
pydantic = "*"
pydantic-settings = "*"
python-dotenv = "*"


[tool.poetry.scripts]
siri_ask_gpt = 'cli:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
